requirements-TF.txt
absl-py==2.3.1
asttokens==3.0.0
astunparse==1.6.3
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
cuda-bindings==13.0.1
cuda-pathfinder==1.2.2
cuda-python==13.0.1
cycler==0.12.1
Cython==3.1.3
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
flatbuffers==25.2.10
fonttools==4.59.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras-nightly==3.12.0.dev2025091003
kiwisolver==1.4.9
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.44.0
lxml==6.0.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.13.0.50
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.3
nvidia-nvjitlink-cu12==12.9.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pandas-ta==0.4.67b0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
ta==0.11.0
TA-Lib==0.6.7
tb-nightly==2.20.0a20250717
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf_nightly==2.21.0.dev20250909
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
